requirements.txt
fastapi[all]
sqlalchemy
alembic
python-dotenv
pydantic-settings
passlib[bcrypt]
python-jose[cryptography]
pandas